// playground-solid/src/App/StyledComponents/MosaicContent/MosaicContent.types.ts
import type { AccessorProps, MosaicItemState } from "@thewaver/ss-components-solid";

export type PageMosaicTileProps = AccessorProps<{
    state: MosaicItemState;
    width: number;
    height: number;
}>;

export type PageMosaicLinkProps = AccessorProps<{
    href: string;
    caption: string;
}>;